backend/requirements.txt
flask==2.2.5
flask-socketio==5.3.0
eventlet==0.33.0
python-socketio==5.5.0
scikit-learn==1.2.2
numpy==1.24.3
aiohttp==3.8.4
pydirectinput==1.1.1
pyjwt==2.6.0
requests==2.28.2